Updated admin page to only show action buttons when appropriate
Added an AdminViewModel class to encapsulate all necessary information for the Admin view. Updated the query for results on the Admin action in the User controller to build the model of AdminViewModels to be shown. Added some logic to the view to determine what buttons should show based on roles.